# Conflicts:
#	reconstruction/cvt/src/main/java/org/jlab/rec/cvt/trajectory/TrajectoryFinder.java conflict merge

+++ reconstruction/cvt/src/main/java/org/jlab/rec/cvt/banks/RecoBankWriter.java
@@ -479,7 +479,7 @@ public class RecoBankWriter {
             for (StateVec stVec : trks.get(i).get_Trajectory()) {
 
                 bank.setShort("id",       k, (short) trks.get(i).get_Id());
-                bank.setShort("detector", k, (byte) stVec.get_SurfaceDetector());
+                bank.setByte("detector",  k, (byte) stVec.get_SurfaceDetector());
                 bank.setByte("sector",    k, (byte) stVec.get_SurfaceSector());
                 bank.setByte("layer",     k, (byte) stVec.get_SurfaceLayer());
                 bank.setFloat("x",        k, (float) (stVec.x()/10.));
@@ -531,7 +531,7 @@ public class RecoBankWriter {
             for (StateVec stVec : trks.get(i).get_Trajectory()) {
 
                 bank.setShort("id",       k, (short) trks.get(i).get_Id());
-                bank.setShort("detector", k, (byte) stVec.get_SurfaceDetector());
+                bank.setByte("detector",  k, (byte) stVec.get_SurfaceDetector());
                 bank.setByte("sector",    k, (byte) stVec.get_SurfaceSector());
                 bank.setByte("layer",     k, (byte) stVec.get_SurfaceLayer());
                 bank.setFloat("x",        k, (float) (stVec.x()/10.));
